+++ backup/moodle2/backup_elang_stepslib.php
@@ -42,7 +42,8 @@ class backup_elang_activity_structure_step extends backup_activity_structure_ste
 
         $elang = new backup_nested_element('elang', ['id'], [
             'name', 'intro', 'introformat', 'language', 'currentversionid', 'grade',
-            'completionfinishattempt', 'jarothreshold', 'options', 'timecreated', 'timemodified',
+            'completionfinishattempt', 'jarothreshold', 'allowtranscriptdownload', 'solutionavailability',
+            'options', 'timecreated', 'timemodified',
             'migrationapproveduserid', 'migrationapprovedtime',
         ]);